very large refactor: PolygonShapeDrawer class has been removed, ShapeDrawer class does filled shapes (not implemented for Sprite Batch yet) BatchManager class that handles batch related calculations two subclasses of FilledShapeDrawer -  one for Batch and another for PolygonBatch DrawerTemplate is parameterised by BatchManager
