+++ drawer/src/space/earlygrey/shapedrawer/AbstractShapeDrawer.java
@@ -4,9 +4,13 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Batch;
+import com.badlogic.gdx.graphics.g2d.PolygonBatch;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.Matrix4;
+
+import space.earlygrey.shapedrawer.FilledPolygonDrawer.BatchFilledPolygonDrawer;
+import space.earlygrey.shapedrawer.FilledPolygonDrawer.PolygonBatchFilledPolygonDrawer;
 
 /**
  * <p>Contains the mechanics for using the Batch and settings such as line width and pixel size.</p>
@@ -20,23 +24,21 @@ public abstract class AbstractShapeDrawer {
     // MEMBERS
     //================================================================================
 
-    private final Batch batch;
-    TextureRegion r;
-    float floatBits;
-    static final int VERTEX_CACHE_SIZE = 2000;
-    final float[] verts;
-    int vertexCount;
-
-    float pixelSize = 1, halfPixelSize = 0.5f * pixelSize;
-    float offset = ShapeUtils.EPSILON * pixelSize;
-    float defaultLineWidth = pixelSize;
+    final BatchManager batchManager;
+    float defaultLineWidth = 1;
     boolean defaultSnap = false;
-    boolean cacheDraws = false;
+
 
     protected static final Matrix4 mat4 = new Matrix4();
 
     // These are named just for clarity
     static final int VERTEX_SIZE = 5, QUAD_PUSH_SIZE = 4 * VERTEX_SIZE;
+
+
+    protected final LineDrawer lineDrawer;
+    protected final PathDrawer pathDrawer;
+    protected final PolygonDrawer polygonDrawer;
+    protected final FilledPolygonDrawer filledPolygonDrawer;
 
 
     //================================================================================
@@ -50,10 +52,19 @@ public abstract class AbstractShapeDrawer {
      */
 
     AbstractShapeDrawer(Batch batch, TextureRegion region) {
-        this.batch = batch;
-        verts = new float[VERTEX_CACHE_SIZE];
-        setTextureRegion(region);
-        setColor(Color.WHITE);
+        if (batch instanceof PolygonBatch) {
+            PolygonBatchManager manager = new PolygonBatchManager((PolygonBatch) batch, region);
+            filledPolygonDrawer = new PolygonBatchFilledPolygonDrawer(manager, this);
+            batchManager = manager;
+        } else {
+            batchManager = new BatchManager(batch, region);
+            filledPolygonDrawer = new BatchFilledPolygonDrawer(batchManager, this);
+        }
+
+        lineDrawer = new LineDrawer(batchManager, this);
+        pathDrawer = new PathDrawer(batchManager, this);
+        polygonDrawer = new PolygonDrawer(batchManager, this);
+
     }
 
 
@@ -111,12 +122,12 @@ public abstract class AbstractShapeDrawer {
      * @return whether drawing joins will likely be discernible
      */
     protected boolean isJoinNecessary(float lineWidth) {
-        return lineWidth > 3 * pixelSize;
+        return lineWidth > 3 * getPixelSize();
     }
 
     protected int estimateSidesRequired(float radiusX, float radiusY) {
         float circumference = (float) (ShapeUtils.PI2 * Math.sqrt((radiusX*radiusX + radiusY*radiusY)/2f));
-        int sides = (int) (circumference / (16 * pixelSize));
+        int sides = (int) (circumference / (16 * getPixelSize()));
         float a = Math.min(radiusX, radiusY), b = Math.max(radiusX, radiusY);
         float eccentricity = (float) Math.sqrt(1-((a*a) / (b*b)));
         sides += (sides * eccentricity) / 16;
@@ -136,11 +147,7 @@ public abstract class AbstractShapeDrawer {
      * @return the previous screen pixel size in world units
      */
     public float setPixelSize(float pixelSize) {
-        float oldPixelSize = this.pixelSize;
-        this.pixelSize = pixelSize;
-        halfPixelSize = 0.5f * pixelSize;
-        offset = ShapeUtils.EPSILON * pixelSize;
-        return oldPixelSize;
+        return batchManager.setPixelSize(pixelSize);
     }
 
     /**
@@ -148,7 +155,7 @@ public abstract class AbstractShapeDrawer {
      * @return the current setting for the pixel size in world units
      */
     public float getPixelSize() {
-        return pixelSize;
+        return batchManager.getPixelSize();
     }
 
     /**
@@ -156,7 +163,7 @@ public abstract class AbstractShapeDrawer {
      * @return the batch this ShapeDrawer was initialised with
      */
     public Batch getBatch() {
-        return batch;
+        return batchManager.getBatch();
     }
 
     /**
@@ -164,7 +171,7 @@ public abstract class AbstractShapeDrawer {
      * @return the current TextureRegion used for drawing
      */
     public TextureRegion getRegion() {
-        return r;
+        return batchManager.getRegion();
     }
 
     /**
@@ -211,13 +218,7 @@ public abstract class AbstractShapeDrawer {
      * @return the previous texture region
      */
     public TextureRegion setTextureRegion(TextureRegion region) {
-        TextureRegion oldRegion = this.r;
-        this.r = region;
-        for (int i = 0; i < verts.length; i+=VERTEX_SIZE) {
-            verts[i + SpriteBatch.U1] = r.getU() + 0.5f*r.getRegionWidth();
-            verts[i + SpriteBatch.V1] = r.getV() + 0.5f*r.getRegionHeight();
-        }
-        return oldRegion;
+        return batchManager.setTextureRegion(region);
     }
 
     /**
@@ -237,9 +238,7 @@ public abstract class AbstractShapeDrawer {
      * @return the previous packed float value of the ShapeDrawer's colour
      */
     public float setColor(float floatBits) {
-        float oldColor = getPackedColor();
-        this.floatBits = floatBits;
-        return oldColor;
+        return batchManager.setColor(floatBits);
     }
 
     /**
@@ -247,114 +246,7 @@ public abstract class AbstractShapeDrawer {
      * @return the packed colour of this ShapeDrawer
      */
     public float getPackedColor() {
-        return floatBits;
+        return batchManager.getPackedColor();
     }
-
-    /**
-     *
-     * @return whether drawing is currently being cached
-     */
-    boolean isCachingDraws() {
-        return cacheDraws;
-    }
-
-    /**
-     * <p>Begin caching draw calls by storing vertex information in a float[] until it all gets set to the
-     * Batch with one call to {@link Batch#draw(Texture, float[], int, int)}.</p>
-     * @return whether drawing was being cached before this method was called
-     */
-    boolean startCaching() {
-        boolean wasCaching = isCachingDraws();
-        this.cacheDraws = true;
-        return wasCaching;
-    }
-
-    /**
-     * <p>Stops caching and calls {@link Batch#draw(Texture, float[], int, int)} if anything is cached.</p>
-     */
-    void endCaching() {
-        this.cacheDraws = false;
-        if (vertexCount>0) pushToBatch();
-    }
-
-
-    //================================================================================
-    // DRAWING METHODS
-    //================================================================================
-
-    void pushVertex() {
-        int i = getVerticesArrayIndex();
-        verts[i + SpriteBatch.C1] = floatBits;
-        vertexCount++;
-    }
-
-    /**
-     * <p>Adds the colour and texture coordinates of four vertices to the cache and progresses the index. If drawing is
-     * not currently being cached, immediately calls {@link #pushToBatch()}.</p>
-     * @return whether the vertex data was pushed to the Batch
-     */
-    void pushQuad() {
-        int i = getVerticesArrayIndex();
-        verts[i + SpriteBatch.C1] = floatBits;
-        verts[i + SpriteBatch.C2] = floatBits;
-        verts[i + SpriteBatch.C3] = floatBits;
-        verts[i + SpriteBatch.C4] = floatBits;
-        vertexCount += 4;
-    }
-
-    /**
-     <p>Adds the colour and texture coordinates of three vertices to the cache and progresses the index. If drawing is
-     * not currently being cached, immediately calls {@link #pushToBatch()}.</p>
-     * @return whether the vertex data was pushed to the Batch
-     */
-    void pushTriangle() {
-        x4(x3());
-        y4(y3());
-        pushQuad();
-    }
-
-    void ensureSpaceForTriangle() {
-        ensureSpace(4);
-    }
-    void ensureSpaceForQuad() {
-        ensureSpace(4);
-    }
-    void ensureSpace(int vertices) {
-        if (verticesRemaining() < vertices) pushToBatch();
-    }
-
-    int verticesRemaining() {
-        return (verts.length - QUAD_PUSH_SIZE * vertexCount) / VERTEX_SIZE;
-    }
-
-    /**
-     * <p>Calls {@link Batch#draw(Texture, float[], int, int)} using the currently cached vertex information.</p>
-     */
-    void pushToBatch() {
-        if (vertexCount == 0) return;
-        batch.draw(r.getTexture(), verts, 0, getVerticesArrayIndex());
-        vertexCount = 0;
-    }
-
-    int getVerticesArrayIndex() {
-        return VERTEX_SIZE * vertexCount;
-    }
-
-    protected void x1(float x1){verts[getVerticesArrayIndex() + SpriteBatch.X1] = x1;}
-    protected void y1(float y1){verts[getVerticesArrayIndex() + SpriteBatch.Y1] = y1;}
-    protected void x2(float x2){verts[getVerticesArrayIndex() + SpriteBatch.X2] = x2;}
-    protected void y2(float y2){verts[getVerticesArrayIndex() + SpriteBatch.Y2] = y2;}
-    protected void x3(float x3){verts[getVerticesArrayIndex() + SpriteBatch.X3] = x3;}
-    protected void y3(float y3){verts[getVerticesArrayIndex() + SpriteBatch.Y3] = y3;}
-    protected void x4(float x4){verts[getVerticesArrayIndex() + SpriteBatch.X4] = x4;}
-    protected void y4(float y4){verts[getVerticesArrayIndex() + SpriteBatch.Y4] = y4;}
-    protected float x1() {return verts[getVerticesArrayIndex() + SpriteBatch.X1];}
-    protected float y1() {return verts[getVerticesArrayIndex() + SpriteBatch.Y1];}
-    protected float x2() {return verts[getVerticesArrayIndex() + SpriteBatch.X2];}
-    protected float y2() {return verts[getVerticesArrayIndex() + SpriteBatch.Y2];}
-    protected float x3() {return verts[getVerticesArrayIndex() + SpriteBatch.X3];}
-    protected float y3() {return verts[getVerticesArrayIndex() + SpriteBatch.Y3];}
-    protected float x4() {return verts[getVerticesArrayIndex() + SpriteBatch.X4];}
-    protected float y4() {return verts[getVerticesArrayIndex() + SpriteBatch.Y4];}
 
 }
