+++ drawer/src/space/earlygrey/shapedrawer/LineDrawer.java
@@ -6,20 +6,20 @@ package space.earlygrey.shapedrawer;
  * @author earlygrey
  */
 
-class LineDrawer extends DrawerTemplate<ShapeDrawer> {
+class LineDrawer extends DrawerTemplate<BatchManager> {
 
-    LineDrawer(ShapeDrawer drawer) {
-        super(drawer);
+    LineDrawer(BatchManager batchManager, AbstractShapeDrawer drawer) {
+        super(batchManager, drawer);
     }
 
     void line(float x1, float y1, float x2, float y2, float lineWidth, boolean snap) {
         pushLine(x1, y1, x2, y2, lineWidth, snap);
-        drawer.pushToBatch();
+        batchManager.pushToBatch();
     }
 
     void pushLine(float x1, float y1, float x2, float y2, float lineWidth, boolean snap) {
 
-        drawer.ensureSpaceForQuad();
+        batchManager.ensureSpaceForQuad();
 
         // dif=(xdif,ydif) is the vector going from (x1, y1) to the first vertex going clockwise around the border of the line.
         // l=(lx,ly) is the vector from (x1, y1) to (x2, y2)
@@ -31,8 +31,8 @@ class LineDrawer extends DrawerTemplate<ShapeDrawer> {
             then extend the line very slightly in each direction, so it covers the centre of the start and end pixels.
             This is because nearest neighbour just takes the colour from right in the centre.
             */
-            float offset = drawer.offset;
-            float pixelSize = drawer.pixelSize, halfPixelSize = drawer.halfPixelSize;
+            float offset = batchManager.offset;
+            float pixelSize = batchManager.pixelSize, halfPixelSize = batchManager.halfPixelSize;
             x1 = ShapeUtils.snap(x1, pixelSize, halfPixelSize) - Math.signum(lx) * offset;
             y1 = ShapeUtils.snap(y1, pixelSize, halfPixelSize) - Math.signum(ly) * offset;
             x2 = ShapeUtils.snap(x2, pixelSize, halfPixelSize) + Math.signum(lx) * offset;
@@ -64,8 +64,8 @@ class LineDrawer extends DrawerTemplate<ShapeDrawer> {
         x4(x2+xdif);
         y4(y2-ydif);
 
-        drawer.pushQuad();
-        if (!drawer.isCachingDraws()) drawer.pushToBatch();
+        batchManager.pushQuad();
+        if (!batchManager.isCachingDraws()) batchManager.pushToBatch();
     }
